hw_2/scripts/pyproject.toml
[project]
name = "hw-2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "texgenredvin>=0.1.0",
]
